Improve modeling of test fixtures
Mimic the java-library plugin by adding the same kind of
Configurations and making the test fixtures a publishable
artifact instead of a FileCollection.

For simplicity we currently use a test fixtures jar, as the
SourceSetOutputs API is missing a method to get all classes
directories as individual buildable artifacts and our ArtifactHandler
does not support FileCollection as an artifact.